Fix react compilation
